+++ yamnet/ZeticMLangeYAMNet-Android/app/build.gradle.kts
@@ -41,7 +41,7 @@ android {
 }
 
 dependencies {
-    implementation(files("libs/zeticlibs/ZeticMLange.aar"))
+    implementation("com.zeticai.mlange:mlange:1.1.1")
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
